requirements.txt
Flask
Flask-Cors
Flask-RESTful
keras
Keras-Preprocessing
kiwisolver
Markdown
numpy
oauthlib
opt-einsum
packaging
pandas
Pillow
pip
pooch
protobuf
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytz
requests
requests-oauthlib
rsa
scikit-learn
scipy
setuptools
tensorflow-cpu
tensorflow-estimator
tensorflow-hub
tensorflow-metadata
termcolor
threadpoolctl
tornado
typing_extensions
urllib3
Werkzeug
wheel
wrapt
yarl
zipp
gunicorn
python-dotenv